+++ src/AppContent.vue
@@ -4,7 +4,7 @@ import { commands } from './bindings.ts'
 import { useMessage, useNotification } from 'naive-ui'
 import LoginDialog from './dialogs/LoginDialog.vue'
 import SearchPane from './panes/SearchPane.vue'
-import ChapterPane from './panes/ChapterPane.vue'
+import ChapterPane from './panes/ChapterPane/ChapterPane.vue'
 import ProgressesPane from './panes/ProgressesPane/ProgressesPane.vue'
 import FavoritePane from './panes/FavoritePane.vue'
 import DownloadedPane from './panes/DownloadedPane/DownloadedPane.vue'
